# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing rental properties, landlords, and tenants",
        "title": "Potom Pridumaem API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Authenticates a user by email or phone and password, and creates a session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Log in a user",
                "parameters": [
                    {
                        "description": "Login data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.Login"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.AuthUser"
                        }
                    },
                    "400": {
                        "description": "invalid request body or credentials",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "description": "Terminates the current user session",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Log out a user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Protected. Returns the full profile of the authenticated user for the current session, including the linked property when the caller is a tenant. Per business requirements this should also trigger an autobilling check; note: autobilling is not implemented yet in this API version, so only profile data is returned for now.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Current session",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Me"
                        }
                    },
                    "401": {
                        "description": "not authenticated",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/auth/profile": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Protected. Partially updates the authenticated user's personal data, including changing the password. Only the provided fields are changed.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Update account settings",
                "parameters": [
                    {
                        "description": "Profile fields to update",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.Profile"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "invalid request body or email already registered",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "401": {
                        "description": "not authenticated",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Creates a new user (landlord or tenant) and immediately logs them in, setting the session cookie",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register a user",
                "parameters": [
                    {
                        "description": "Registration data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.Register"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.AuthUser"
                        }
                    },
                    "400": {
                        "description": "invalid request body, invalid role, or email already registered",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.Login": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "ivan.petrov@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "strongPass123"
                }
            }
        },
        "request.Profile": {
            "type": "object",
            "properties": {
                "document": {
                    "type": "string",
                    "example": "4510 123456"
                },
                "email": {
                    "type": "string",
                    "example": "newemail@example.com"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2,
                    "example": "Ivan Kolesnikov"
                },
                "password": {
                    "type": "string",
                    "minLength": 6,
                    "example": "newSuperSecurePassword"
                },
                "paymentCard": {
                    "type": "string",
                    "example": "1111222233334444"
                },
                "phone": {
                    "type": "string",
                    "example": "+79997776655"
                }
            }
        },
...
        "request.Register": {
            "type": "object",
            "required": [
                "email",
                "initialRole",
                "name",
                "password",
                "phone"
            ],
            "properties": {
                "document": {
                    "type": "string",
                    "example": "4510 123456"
                },
                "email": {
                    "type": "string",
                    "example": "ivan.petrov@example.com"
                },
                "initialRole": {
                    "type": "string",
                    "enum": [
                        "landlord",
                        "tenant"
                    ],
                    "example": "landlord"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2,
                    "example": "Ivan Petrov"
                },
                "password": {
                    "type": "string",
                    "minLength": 6,
                    "example": "strongPass123"
                },
                "paymentCard": {
                    "type": "string",
                    "example": "1234567812345678"
                },
                "phone": {
                    "type": "string",
                    "example": "+79161234567"
                }
            }
        },
        "response.AuthUser": {
            "description": "Краткая информация о пользователе, возвращаемая при регистрации и входе.",
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "ivanov@example.com"
                },
                "id": {
                    "type": "string",
                    "example": "user-a9b8c7d"
                },
                "name": {
                    "type": "string",
                    "example": "Иванов Иван Иванович"
                },
                "role": {
                    "type": "string",
                    "example": "tenant"
                }
            }
        },
        "response.Error": {
            "description": "Стандартный формат ошибки API",
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid request body"
                }
            }
        },
        "response.Me": {
            "description": "Полный профиль авторизованного пользователя текущей сессии.",
            "type": "object",
            "properties": {
                "document": {
                    "type": "string",
                    "example": "Паспорт РФ 4512 № 345678"
                },
                "email": {
                    "type": "string",
                    "example": "ivanov@example.com"
                },
                "id": {
                    "type": "string",
                    "example": "user-a9b8c7d"
                },
                "name": {
                    "type": "string",
                    "example": "Иванов Иван Иванович"
                },
                "paymentCard": {
                    "type": "string",
                    "example": "4276111122223333"
                },
                "phone": {
                    "type": "string",
                    "example": "+79991112233"
                },
                "role": {
                    "type": "string",
                    "example": "tenant"
                },
                "tenantPropertyId": {
                    "type": "string",
                    "example": "prop-z8y7x6w"
                }
            }
        }
    },
    "securityDefinitions": {
        "CookieAuth": {
            "description": "Session cookie set on successful login",
            "type": "apiKey",
            "name": "session_id",
            "in": "cookie"
        }
    }
}